Separate out the Composer store
